finally all tests works extracted main outside so google test does not complain about two mains
