+++ CMakeLists.txt
@@ -5,7 +5,5 @@ project(MiscadaProjectGp12Tests)
 set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
-enable_testing()
-
 add_subdirectory(src)
 add_subdirectory(tests)
