+++ tests/CMakeLists.txt
@@ -10,6 +10,8 @@ FetchContent_MakeAvailable(googletest)
 add_library(GTest::GTest INTERFACE IMPORTED)
 target_link_libraries(GTest::GTest INTERFACE gtest_main)
 
+enable_testing()
+
 add_executable(
   gs_test
   gs_test.cpp
@@ -21,5 +23,5 @@ target_link_libraries(
   GTest::GTest
   gs
 )
-
-add_test(gs_gtests gs_test)
+include(GoogleTest)
+gtest_discover_tests(gs_test)
